### 2018-07-13 #822 @whedon
[PRE REVIEW]: pfla: A Python Package for Dental Facial Analysis using Computer Vision and Statistical Shape Analysis
**Submitting author:** @maxrousseau (<a href="http://orcid.org/0000-0002-1417-2511">Maxime Rousseau</a>)
**Repository:** <a href="https://github.com/maxrousseau/pfla" target ="_blank">https://github.com/maxrousseau/pfla</a>
**Version:** v0.1.4
