### 2018-07-16 #822 @pjotrp
This is more of an AI thing - which editor do we have specialized in this direction?
