### 2018-07-16 #822 @maxrousseau
Indeed, maybe I should suggest Ariel Rokem @arokem as an editor.
