### 2018-07-18 #822 @maxrousseau
@arfon - Do you know of any editor that would be available for this type of paper?

Thank you,

Maxime Rousseau
