### 2018-07-20 #822 @arokem
👋 @mnarayan: would you be willing to review this article for JOSS?
