### 2018-07-20 8afa3a1
removed and replaced lena.jpg from paper, test and readme
