### 2018-07-20 9d1aa55
modified manifest.in and fixed packaging
