### 2018-07-27 #822 @arokem
Hey @glemaitre! Would you be willing to review this article for [The Journal of Open Source Software](http://joss.theoj.org/)? It seems to be within your area of expertise. To read more about what the review process entails, please see [this page](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html)
