### 2018-07-28 #822 @glemaitre
I will be able to review it as well

Sent from my phone - sorry to be brief and potential misspell.
