### 2018-08-20 maxrousseau/pfla#4 @glemaitre
JOSS - Problem of install with conda
I cannot install the package out-of-the-box on linux using conda:

```
(base) ✘-2 /tmp
16:45 $ conda create --name pfla python=3
Solving environment: done

## Package Plan ##

  environment location: /home/lemaitre/miniconda3/envs/pfla

  added / updated specs:
    - python=3

The following packages will be downloaded:

    package                    |            build
    ---------------------------|-----------------
    libgcc-ng-8.2.0            |       hdf63c60_0         7.6 MB
    certifi-2018.8.13          |           py37_0         139 KB
    setuptools-40.0.0          |           py37_0         551 KB
    libstdcxx-ng-8.2.0         |       hdf63c60_0         2.9 MB
    ------------------------------------------------------------
                                           Total:        11.2 MB

The following NEW packages will be INSTALLED:

    ca-certificates: 2018.03.07-0
    certifi:         2018.8.13-py37_0
    libedit:         3.1.20170329-h6b74fdf_2
    libffi:          3.2.1-hd88cf55_4
    libgcc-ng:       8.2.0-hdf63c60_0
    libstdcxx-ng:    8.2.0-hdf63c60_0
    ncurses:         6.1-hf484d3e_0
    openssl:         1.0.2p-h14c3975_0
    pip:             10.0.1-py37_0
    python:          3.7.0-hc3d631a_0
    readline:        7.0-ha6073c6_4
    setuptools:      40.0.0-py37_0
    sqlite:          3.24.0-h84994c4_0
    tk:              8.6.7-hc745277_3
    wheel:           0.31.1-py37_0
    xz:              5.2.4-h14c3975_4
    zlib:            1.2.11-ha838bed_2

Proceed ([y]/n)? y

Downloading and Extracting Packages
libgcc-ng-8.2.0      | 7.6 MB    | ########################################## | 100%
certifi-2018.8.13    | 139 KB    | ########################################## | 100%
setuptools-40.0.0    | 551 KB    | ########################################## | 100%
libstdcxx-ng-8.2.0   | 2.9 MB    | ########################################## | 100%
Preparing transaction: done
Verifying transaction: done
Executing transaction: done
#
# To activate this environment, use
#
#     $ conda activate pfla
#
# To deactivate an active environment, use
#
#     $ conda deactivate

(base) ✔ /tmp
16:46 $ conda activate pfla
(pfla) ✔ /tmp
16:46 $ pip install pfla
Collecting pfla
  Downloading https://files.pythonhosted.org/packages/bb/8b/af02451e1cc9d9521d8ee3f81818d8f8dec48b1683c32853d9d265f4ac02/pfla-0.1.6-py3-none-any.whl (1.9MB)
    100% |████████████████████████████████| 1.9MB 4.1MB/s
Collecting dlib (from pfla)
  Downloading https://files.pythonhosted.org/packages/df/aa/6a9bb2a763107bb2606d6ee1aa65fcd3b51375a9ef6436e9c9280b0dd63c/dlib-19.15.0.tar.gz (3.3MB)
    100% |████████████████████████████████| 3.3MB 3.3MB/s
Collecting progress (from pfla)
  Downloading https://files.pythonhosted.org/packages/e9/ff/7871f3736dc6707435b2a2f217c46b5a5bc6ea7e0a9a443cd69146a1afd1/progress-1.4.tar.gz
Collecting opencv-python (from pfla)
  Downloading https://files.pythonhosted.org/packages/8f/8f/a5d2fa3a3309c4e4aa28eb989d81a95b57c63406b4d439758a1a0a810c77/opencv_python-3.4.2.17-cp37-cp37m-manylinux1_x86_64.whl (25.0MB)
    100% |████████████████████████████████| 25.0MB 980kB/s
Collecting rpy2 (from pfla)
  Downloading https://files.pythonhosted.org/packages/f1/98/c7652cc9d7fc0afce74d2c30a52b9c9ac391713a63d037e4ab8feb56c530/rpy2-2.9.4.tar.gz (194kB)
    100% |████████████████████████████████| 204kB 7.5MB/s
    Complete output from command python setup.py egg_info:
    Error: Tried to guess R's HOME but no command 'R' in the PATH.

Command "python setup.py egg_info" failed with error code 1 in /tmp/pip-install-8cv6gq93/rpy2/

```
