### 2018-08-21 2cf0725
modified installation instructions and added to documentation
