### 2018-08-21 maxrousseau/pfla#4 @maxrousseau
This error is related to pip being unable to find R on your computer to install the required rpy2 package.
Do you have R installed on your machine?
edit: spelling
