### 2018-08-21 9e0e100
travis changed command to python 3.5.6
