### 2018-08-21 maxrousseau/pfla#4 @glemaitre
Nop. It seems that you only have a single file requiring R: https://github.com/maxrousseau/pfla/blob/master/pfla/fcn/stats.R

IMO, it would make a lot of sense to convert this file to Python. Then, you package will be a pure Python package and will be easy to install using pip.
