### 2018-08-21 maxrousseau/pfla#4 @maxrousseau
Indeed there is only one file in R but I cannot rewrite it in python since it is using this R script that the statistical shape analysis is conducted. To do so I had to use the "shapes" package which implements GPA, shapePCA and statistical testing of the group shapes.
