### 2018-08-22 8f32640
github release badge fix
