### 2018-08-23 maxrousseau/pfla#4 @maxrousseau
@glemaitre can you confirm to me that the problem regarding the rpy2 package was resolved so that I can close this issue.
Thank you.
