### 2018-08-23 maxrousseau/pfla#4 @glemaitre
I did not have time to come back on the review yet.

On Thu, 23 Aug 2018 at 19:32, Maxime Rousseau <notifications@github.com>
wrote:

> @glemaitre <https://github.com/glemaitre> can you confirm to me that the
> problem regarding the rpy2 package was resolved so that I can close this
> issue.
> Thank you.
>
