### 2018-09-02 #855 @arokem
@glemaitre : have you been able to resolve the installation issue mentioned in https://github.com/maxrousseau/pfla/issues/4?

@mnarayan : have you had a chance to take a look at this?
