### 2018-09-07 #855 @mnarayan
I have looked at the repository and it still needs some work. Will finish submitting all the issues around documentation this weekend
