### 2018-09-08 43de4d7
paper improvements
