+++ paper/paper.bib
@@ -1,5 +1,5 @@
 @article{dryden92,
-author = {DRYDEN, I. L. and MARDIA, K. V.},
+author = {Dryden, I. L. and Mardia, K. V.},
 title = {Size and shape analysis of landmark data},
 journal = {Biometrika},
 volume = {79},
@@ -9,6 +9,55 @@ year = {1992},
 doi = {10.1093/biomet/79.1.57},
 URL = {http://dx.doi.org/10.1093/biomet/79.1.57},
 eprint = {/oup/backfile/content_public/journal/biomet/79/1/10.1093/biomet/79.1.57/2/79-1-57.pdf}
+}
+
+
+@book{dryden16,
+  title={Statistical shape analysis: with applications in R},
+  author={Dryden, Ian L and Mardia, Kanti V},
+  year={2016},
+  publisher={John Wiley \& Sons}
+}
+
+@article{goodall91,
+  title={Procrustes methods in the statistical analysis of shape},
+  author={Goodall, Colin},
+  journal={Journal of the Royal Statistical Society. Series B (Methodological)},
+  pages={285--339},
+  year={1991},
+  publisher={JSTOR}
+}
+
+@article{gower75,
+  title={Generalized procrustes analysis},
+  author={Gower, John C},
+  journal={Psychometrika},
+  volume={40},
+  number={1},
+  pages={33--51},
+  year={1975},
+  publisher={Springer}
+}
+
+@article{kent94,
+  title={The complex Bingham distribution and shape analysis},
+  author={Kent, John T},
+  journal={Journal of the Royal Statistical Society. Series B (Methodological)},
+  pages={285--299},
+  year={1994},
+  publisher={JSTOR}
+}
+
+
+@article{ten77,
+  title={Orthogonal Procrustes rotation for two or more matrices},
+  author={Ten Berge, Jos MF},
+  journal={Psychometrika},
+  volume={42},
+  number={2},
+  pages={267--276},
+  year={1977},
+  publisher={Springer}
 }
 
 @INPROCEEDINGS{viola01,
@@ -40,7 +89,7 @@ month={},}
 	            address      = {Vienna, Austria},
 		         year         = 2017,
 			      note         = {Contributed package, Version 1.2.3.},
-			           url          = {http://www.R-project.org}
+			           url          ={https://cran.r-project.org/web/packages/shapes/shapes.pdf}
 				      }
 
 @INPROCEEDINGS{redmon16,
