+++ paper/paper.md
@@ -24,17 +24,26 @@ date: 17 May 2018
 bibliography: paper.bib
 
 ---
-# Summary
+# Introduction
+
 This paper outlines the workings of the software used to conduct automatic
 facial analyses on patients of the BBDC 7701 protocol investigating the natural
-history of Osteogenesis Imperfecta.
+history of Osteogenesis Imperfecta (OI).
+
+The software takes two image directories containing facial pictures of the study
+subjects as inputs. Each image is processed through face detection and landmark
+allocation producing a matrix of coordinates for each study group. Theses
+matrices are then used for the statistical comparison of the groups. The
+results from the analysis is printed to the terminal (Figure 1). 
+
+![Schematic representation of the pfla software for OI facial
+analysis](./pfla.png)
 
 One of the main advantages of using this software for clinical image analysis
 is that it automatically places landmarks, standardizing and expediting
-measurement procedures. The program was  written in python and R [@r]
-programming languages using OpenCV [@opencv] and Dlib [@dlib09]
-libraries, as well as the publicly available facial annotation tool by
-[@sagonas13].
+measurement procedures. This task was previously conducted manually which was
+very time consuming and prone to operator error and bias. With this novel
+approach we are able to circumvent these issues allowing us to have an efficient and objective way of conducting facial analysis.
 
 
 This paper simultaneously presents a broad framework for facial analysis, while
@@ -42,6 +51,14 @@ describing a specific package applying trained models and statistical analysis.
 These algorithms can be interchanged (i.e. YOLO [@redmon16]) and the R
 script modified to suit the needs of a particular study.
 
+
+
+# Software
+
+The program was  written in python and R [@r]
+programming languages using OpenCV [@opencv] and Dlib [@dlib09]
+libraries, as well as the publicly available facial annotation tool by
+[@sagonas13].
 
 The program takes as input two folders of dental anteroposterior .jpg images,
 before assigning landmarks to each object of interest. All coordinates are
@@ -66,6 +83,8 @@ coordinates of the rectangles from the detected faces in each image. The ldmks/
 directory contains the matrices of landmarks for each group to be analyzed
 using the R script.
 
+## Image Processing
+
 After initial preparation, images go through a Haar Cascade classifier trained
 to detect faces [@viola01]. This algorithm scans the input through
 the scope of a small rectangle. It sums up the mean features thus detected,
@@ -80,20 +99,24 @@ of disease by comparing groups of patients with clinical conditions to controls.
 
 The antero posterior analysis consists of $l=68$ landmarks automatically placed
 on patient photographs via software. These sets of coordinates produce matrices
-of $k=2$ dimensions. The matrices are represented as such:
+of $k=2$ dimensions. The matrix for a single study group are represented as such:
 
 \begin{center}
-$M_{patient}=[x_1,x_2,...,x_l,y_1,y_2,...,y_l]$
+$m^{[s]}=[x_1,x_2,...,x_{68},y_1,y_2,...,y_{68}]$
 \end{center}
-Where $l$ represents the number of points attached to a photographs.
+
+Where the $s$ superscript represents the subject attached to this row of
+coordinates. 
+
+## Statistical Shape Analysis
 
 Statistical shape analysis has mostly been used in the field of evolutionary
-biology for the analysis of skeletal artifacts. It also has applications in the
+biology for the analysis of skeletal artifacts [@dryden-shapes]. It also has applications in the
 medical fields, most notably in imaging analysis. The pfla package performs the
-image processing described above in order to conduct statistical analysis. The
-R script uses the "shapes" package by [@dryden-shapes]. First, the various
[…]
