### 2018-09-09 89c7166
contribution guidelines minor improvement
