### 2018-09-15 maxrousseau/pfla#4 @glemaitre
So you should probably add a `environment.yml` file for conda user. Conda allows for installing R and Python packages at once. Something like this should work:

```
name: pfla
channels:
  - conda-forge
dependencies:
  - numpy
  - pandas
  - opencv
  - dlib
  - progress
  - rpy2
  - r-base
  - r-shape
  - r-foreach
  - pip:
    - imutils
```

The version will be resolved by conda as well.
