### 2018-09-16 maxrousseau/pfla#4 @maxrousseau
Unfortunately the "shapes" (not to be confused with "shape") package is not available on conda. That is why i dont recommend using conda for this and instead something like VirtualEnv.

Edit:
I was able to install it with the following commands
```sh
conda install conda-build
conda skeleton cran --recursive shapes
conda build r-shapes
conda install -c local r-shapes
```
**important to note that default R installations will cause an issue at build, to go around this temporarily remove your lib paths from the .Renviron or .Rprofile files.
