### 2018-09-16 #855 @glemaitre
I open 2 issues which need to be addressed (for unit testing and file fetching).
I need to open another one but if those could be addressed first, it could be great.
