### 2018-09-16 b019e02
added conda env and requirements file
