+++ setup.py
@@ -9,9 +9,9 @@ setup(name='pfla',
       description='Python facial landmarking and analysis',
       long_description=readme(),
       classifiers=[
-	'Programming Language :: Python :: 3.5',
-	'Topic :: Scientific/Engineering :: Medical Science Apps.',
-	'Topic :: Scientific/Engineering :: Image Recognition'
+          'Programming Language :: Python :: 3.5',
+          'Topic :: Scientific/Engineering :: Medical Science Apps.',
+          'Topic :: Scientific/Engineering :: Image Recognition'
 	],
       url='https://github.com/maxrousseau/pfla',
       author='Maxime Rousseau',
@@ -20,6 +20,4 @@ setup(name='pfla',
       include_package_data=True,
       license='MIT',
       packages=['pfla'],
-      install_requires=['opencv-python', 'dlib','imutils','numpy','argparse','pandas','rpy2',
-                       'progress'],
       zip_safe=False)
