+++ README.md
@@ -47,10 +47,10 @@ Features
 Requirements and Dependencies
 -----------------------------
 
--   Python 3.5
+-   Python 3.5 (or higher)
 -   Python packages: opencv-python, dlib, imutils, numpy, argparse, pandas, rpy2, progress
 -   Linux operating system
--   R 3.3 (or more recent)
+-   R 3.3 (or more or higher)
 -   R packages: shapes, foreach
 
 Installation
@@ -60,12 +60,29 @@ Important: in order for the required package rpy2 to install sucessfully, you
 will need to have R version 3.3 or higher as well as the packages 'shapes' and
 'foreach'
 
-To install enter the following commands:
+To install with **conda**:
 
 ```shell
-$ pip install pfla
+conda env create -f environment.yml
+source activate pfla
+mkdir shapes
+cd shapes
+conda skeleton cran --recursive shapes
+conda build r-shapes
+conda install -c local r-shapes
+pip install pfla
 ```
 
+To install with **pip**:
+
+```shell
+pip install -r requirements.txt
+pip install pfla
+```
+Then in R
+```R
+install.packages("shapes", "foreach")
+```
 
 Additionnal steps, the 68 landmark dat file is too large for pip packaging.
 You can download it [here](pfla/data/shape_predictor_68_face_landmarks.dat).
