### 2018-09-16 maxrousseau/pfla#4 @maxrousseau
Conda environment.yml and pip requirements.txt were added as suggested. (the install_requires was also removed). Let me know if this issue can be closed.
