+++ setup.py
@@ -5,7 +5,7 @@ def readme():
         return f.read()
 
 setup(name='pfla',
-      version='0.1.6',
+      version='0.1.7',
       description='Python facial landmarking and analysis',
       long_description=readme(),
       classifiers=[
