+++ docs/source/install.rst
@@ -17,14 +17,22 @@ Important: in order for the required package rpy2 to install sucessfully, you
 will need to have R version 3.3 or higher as well as the packages 'shapes' and
 'foreach'
 
-To install enter the following commands:
+To install with **conda**:
 
-        $ pip install pfla
+.. code-block:: shell
+  
+  conda env create -f environment.yml
+  source activate pfla
+  mkdir shapes
+  cd shapes
+  conda skeleton cran --recursive shapes
+  conda build r-shapes
+  conda install -c local r-shapes
+  pip install pfla
 
+To install with pip:
 
-Additionnal steps, the 68 landmark dat file is too large for pip packaging.
-You can download it [here](pfla/data/shape_predictor_68_face_landmarks.dat).
+.. code-block:: shell
 
-Place the downloaded dat file in the following directory:
-
-        $ ~/.local/lib/python3.5/site-packages/pfla/data/
+  pip install -r requirements.txt
+  pip install pfla
