### 2018-09-17 beb0f5c
fetcher fix and preparation for repackaging
