### 2018-09-17 a9616db
rebuild doc readme install requs
