### 2018-09-17 a269b8b
fix requirements for docs build
