+++ rtd_requirements.txt
@@ -1,0 +1,7 @@
+dlib
+glob2
+imutils
+numpy
+opencv-python
+pandas
+progress
