+++ .travis.yml
@@ -16,6 +16,7 @@ before_install:
         - gpg -a --export E084DAB9 | sudo apt-key add -
         - sudo apt-get update
         - sudo apt-get install r-base r-base-dev
+        - pip3 install -r requirements.txt
 
 install: pip3 install pfla
 
