+++ pfla/fcn/annotate.py
@@ -5,6 +5,7 @@
 # with 68 landmarks and output the resulting image and its coordinate matrix
 #
 #------------------------------------------------------------------------------
+import os
 import sys
 import cv2
 import dlib
@@ -12,14 +13,9 @@ import numpy as np
 import csv
 import pandas as pd
 
-for p in sys.path:
-    if 'packages' in p:
-        mod_path = p
-mod_path = mod_path + '/pfla'
-
 class FaceAnnotator(object):
 
-    def __init__(self, img_id, classifier):
+    def __init__(self, img_id, classifier, mod_path):
         """
         Fit the 68 landmarks to the detected face.
 
@@ -31,6 +27,8 @@ class FaceAnnotator(object):
             Identification number of the image being processed
         classifier : string
             Path to the classifier xml file containing 68 landmark algorithm
+        mod_path : string
+            Path to the pfla module
 
         Returns
         -------
@@ -38,6 +36,7 @@ class FaceAnnotator(object):
         """
         self.img_id = img_id
         self.classifier = classifier
+        self.mod_path = mod_path
 
     def run_annotator(self):
         """
@@ -59,7 +58,10 @@ class FaceAnnotator(object):
             detected face.
         """
         rect = []
-        with open(mod_path + "/data/faces/" + self.img_id + ".csv", newline="") as csvfile:
+        with open(
+            os.path.join(self.mod_path, "data/faces/", str(self.img_id + ".csv")),
+            newline=""
+            ) as csvfile:
             face_reader = csv.reader(csvfile)
             for row in face_reader:
                 rect.append(row)
@@ -74,7 +76,8 @@ class FaceAnnotator(object):
         predictor = dlib.shape_predictor(self.classifier)
 
         # place landmarks and convert to pandas dataframe
-        img_prep = cv2.imread(mod_path + "/img/img_prep/" + self.img_id + ".jpg")
+        img_prep = cv2.imread(os.path.join(self.mod_path, "img/img_prep/",
+                                           str(self.img_id + ".jpg")))
         landmarks = predictor(img_prep, dlib_rect).parts()
         landmarks_np = np.matrix([[p.x, p.y] for p in landmarks])
         dataframe = pd.DataFrame(landmarks_np)
@@ -94,7 +97,7 @@ class FaceAnnotator(object):
             )
             cv2.circle(img_proc, position, 1, color=(255, 0, 0))
         cv2.rectangle(img_proc, (x1, y1), (x2, y2), (0,255,0), 2)
-        cv2.imwrite(mod_path + "/img/img_proc/" + self.img_id + ".jpg", img_proc)
+        cv2.imwrite(os.path.join(self.mod_path, "img/img_proc/", str(self.img_id + ".jpg")), img_proc)
 
 
         return dataframe
