+++ paper/paper.bib
@@ -36,7 +36,8 @@ eprint = {/oup/backfile/content_public/journal/biomet/79/1/10.1093/biomet/79.1.5
   number={1},
   pages={33--51},
   year={1975},
-  publisher={Springer}
+  publisher={Springer},
+  doi = {https://doi.org/10.1007/BF02291478}
 }
 
 @article{kent94,
@@ -45,7 +46,7 @@ eprint = {/oup/backfile/content_public/journal/biomet/79/1/10.1093/biomet/79.1.5
   journal={Journal of the Royal Statistical Society. Series B (Methodological)},
   pages={285--299},
   year={1994},
-  publisher={JSTOR}
+  publisher={JSTOR},
 }
 
 
@@ -57,7 +58,8 @@ eprint = {/oup/backfile/content_public/journal/biomet/79/1/10.1093/biomet/79.1.5
   number={2},
   pages={267--276},
   year={1977},
-  publisher={Springer}
+  publisher={Springer},
+  doi = {https://doi.org/10.1007/BF02294053}
 }
 
 @INPROCEEDINGS{viola01,
@@ -79,7 +81,7 @@ month={},}
   journal = {Journal of Machine Learning Research},
   year = {2009},
   volume = {10},
-  pages = {1755-1758},
+  pages = {1755-1758}
 }
 
 @Manual{dryden-shapes,
@@ -153,6 +155,7 @@ author = "Chiara Brombin and Luigi Salmaso"
          Journal="Lancet",
 	    Year="2004",
 	       Volume="363",
+	       doi = {https://doi.org/10.1016/S0140-6736(04)16051-0},
 	          Number="9418",
 		     Pages="1377--1385",
 		        Month="Apr"
@@ -164,6 +167,7 @@ author = "Chiara Brombin and Luigi Salmaso"
       Title="{{B}rachycephalic, dolichocephalic and mesocephalic: {I}s it appropriate to describe the face using skull patterns?}",
          Journal="Dental Press J Orthod",
 	    Year="2013",
+		doi = {http://dx.doi.org/10.1590/S2176-94512013000300025},
 	       Volume="18",
 	          Number="3",
 		     Pages="159--163"
@@ -178,7 +182,8 @@ author = "Chiara Brombin and Luigi Salmaso"
 	       Volume="1",
 	          Number="8072",
 		     Pages="1041--1042",
-		        Month="May"
+		        Month="May",
+   doi = {https://doi.org/10.1007/s10354-015-0368-3}
 			}
 
 
@@ -189,6 +194,7 @@ author = "Chiara Brombin and Luigi Salmaso"
 	volume = "47",
 	pages = "3-18",
 	year = "2016",
+   doi = {https://doi.org/10.1016/j.imavis.2016.01.002}
 }
 
 @article{sagonas13,
