### 2018-10-01 63ee70b
example dataset and bug fixes
