### 2018-10-22 #855 @arokem
@mnarayan : I see that all your issues on the software repo are closed. Does that mean that the software now meets all the criteria in your estimation?

@glemaitre : have you had a chance to take a look whether the issues you raised were properly addressed?
