### 2018-11-11 #855 @mnarayan
I am satisfied with all the corrections and have no more issues to raise.
