### 2018-11-19 #855 @maxrousseau
@glemaitre: Any final thoughts?
