### 2018-11-29 #855 @arokem
OK. Since this review has gone on for a while, one reviewer has already provided a +1, and since the author has addressed the issues raised by the remaining reviewer, I'm going to call it.

@mnarayan, @glemaitre : thanks for your reviews!

@maxrousseau : your article is ready to be accepted by JOSS.

To proceed, please create an archive of the accepted version of the software (e.g., using [zenodo](https://zenodo.org)) and post the DOI here for archiving.
