### 2018-11-29 bc3ea17
COSMIT: PEP8 and simplification of some python code
