+++ setup.py
@@ -10,6 +10,8 @@ setup(name='pfla',
       long_description=readme(),
       classifiers=[
           'Programming Language :: Python :: 3.5',
+          'Programming Language :: Python :: 3.6',
+          'Programming Language :: Python :: 3.7',
           'Topic :: Scientific/Engineering :: Medical Science Apps.',
           'Topic :: Scientific/Engineering :: Image Recognition'
 	],
