+++ pfla/__init__.py
@@ -1,25 +1,27 @@
 # -*- coding: utf-8 -*-
-from pfla.fcn import img_prep
-from pfla.fcn import face_detect
-from pfla.fcn import annotate
-from pfla.fcn import analyze
-from pfla.fcn import fetcher
-from progress.bar import IncrementalBar
-from pathlib import Path
+import argparse
+import csv
+import glob
+import os
 import shutil
 import sys
-import os
-import argparse
-import cv2
-import glob
-import pandas as pd
-import numpy as np
-import csv
 
+import cv2
+import numpy as np
+import pandas as pd
+
+from pathlib import Path
+from progress.bar import IncrementalBar
+
+from .fcn import img_prep
+from .fcn import face_detect
+from .fcn import annotate
+from .fcn import analyze
+from .fcn import fetcher
 
 mod_path = os.path.dirname(os.path.abspath(__file__))
-shape_pred = Path(os.path.join(mod_path,
-                          "data/shape_predictor_68_face_landmarks.dat"))
+shape_pred = Path(os.path.join(
+    mod_path, "data", "shape_predictor_68_face_landmarks.dat"))
 
 if shape_pred.is_file():
     pass
@@ -47,10 +49,10 @@ groups = ["g1", "g2"]
 
 for group in groups:
     # create directories for storage
-    img_raw_d = (os.path.join(mod_path, "img/img_raw/"))
-    img_prep_d = (os.path.join(mod_path, "img/img_prep/"))
-    img_proc_d = (os.path.join(mod_path, "img/img_proc/"))
-    data_d = (os.path.join(mod_path, "data/faces/"))
+    img_raw_d = (os.path.join(mod_path, "img", "img_raw"))
+    img_prep_d = (os.path.join(mod_path, "img", "img_prep"))
+    img_proc_d = (os.path.join(mod_path, "img", "img_proc"))
+    data_d = (os.path.join(mod_path, "data", "faces"))
     list_dir = [img_raw_d, img_prep_d, img_proc_d, data_d]
     for direc in list_dir:
         if os.path.isdir(os.path.join(direc, group)):
@@ -59,32 +61,33 @@ for group in groups:
             os.mkdir(os.path.join(direc, group))
 
     # clear directories before running functions
-    img_raw_gd = (os.path.join(mod_path, "img/img_raw/", group, "*"))
-    img_prep_gd = (os.path.join(mod_path, "img/img_prep/", group, "*"))
-    img_proc_gd = (os.path.join(mod_path, "img/img_proc/", group, "*"))
-    data_gd = (os.path.join(mod_path, "data/faces/", group, "*"))
+    img_raw_gd = (os.path.join(mod_path, "img", "img_raw", group, "*"))
+    img_prep_gd = (os.path.join(mod_path, "img", "img_prep", group, "*"))
+    img_proc_gd = (os.path.join(mod_path, "img", "img_proc", group, "*"))
+    data_gd = (os.path.join(mod_path, "data", "faces", group, "*"))
     list_group_dir = [img_raw_gd, img_prep_gd, img_proc_gd, data_gd]
     for direc in list_group_dir:
         direc = glob.glob(direc)
         for f in direc:
             os.remove(f)
 
+
 def img_processing(img_id):
     """preparation, face detection and landamarking"""
-    img_id = str(img_id)
 
     # begin by creating image object and preparing for processing
     img = img_prep.RawImage(
-        os.path.join(mod_path, "img/img_raw/", str(img_id + ".jpg")),
-        os.path.join(mod_path, "img/img_prep/", str(img_id + ".jpg")),
+        os.path.join(mod_path, "img", "img_raw", "{}.jpg".format(img_id)),
[…]
