+++ pfla/test.py
@@ -2,18 +2,19 @@
 import os
 import filecmp
 import unittest
-from fcn import img_prep
-from fcn import face_detect
-from fcn import annotate
+from pfla.fcn import img_prep
+from pfla.fcn import face_detect
+from pfla.fcn import annotate
 import sys
 
 mod_path = os.path.dirname(os.path.abspath(__file__))
 
-dirs = ["img_raw/", "img_prep/", "img_proc/"]
+dirs = ["img_raw", "img_prep", "img_proc"]
 for di in dirs:
-    path = os.path.join(mod_path, "img/", di)
+    path = os.path.join(mod_path, "img", di)
     if not os.path.exists(path):
         os.makedirs(path)
+
 
 class TestPfla(unittest.TestCase):
     """tests for the pfla package"""
@@ -21,16 +22,17 @@ class TestPfla(unittest.TestCase):
     def setUp(self):
         """load image, prepare and save in test directory"""
         self.test_raw = img_prep.RawImage(
-            os.path.join(mod_path, "test/testpic.jpg"),
-            os.path.join(mod_path, "img/img_prep/00_test.jpg"),
+            os.path.join(mod_path, "test", "testpic.jpg"),
+            os.path.join(mod_path, "img", "img_prep", "00_test.jpg"),
             "00_test"
         )
         self.test_prep = self.test_raw.prepare()
         self.fdetector = face_detect.FaceDectector(
-            os.path.join(mod_path, "img/img_prep/00_test.jpg"),
-            os.path.join(mod_path, "data/haarcascade_frontalface_default.xml"),
-            (100,100),
-            (500,500),
+            os.path.join(mod_path, "img", "img_prep", "00_test.jpg"),
+            os.path.join(mod_path, "data",
+                         "haarcascade_frontalface_default.xml"),
+            (100, 100),
+            (500, 500),
             "00_test",
             mod_path
         )
@@ -39,7 +41,7 @@ class TestPfla(unittest.TestCase):
         self.fannotator = annotate.FaceAnnotator(
             "00_test",
             os.path.join(mod_path,
-                         "data/shape_predictor_68_face_landmarks.dat"),
+                         "data", "shape_predictor_68_face_landmarks.dat"),
             mod_path
         )
         self.mat = self.fannotator.run_annotator()
@@ -47,8 +49,8 @@ class TestPfla(unittest.TestCase):
     def test01_img_prep(self):
         """test image preparation function"""
         success = filecmp.cmp(
-            os.path.join(mod_path, "img/img_prep/00_test.jpg"),
-            os.path.join(mod_path, "test/testpic_gray.jpg"),
+            os.path.join(mod_path, "img", "img_prep", "00_test.jpg"),
+            os.path.join(mod_path, "test", "testpic_gray.jpg"),
             shallow=False
         )
         self.assertTrue(success)
@@ -56,8 +58,8 @@ class TestPfla(unittest.TestCase):
     def test02_face_processing(self):
         """test face processing functions"""
         success = filecmp.cmp(
-            os.path.join(mod_path, "img/img_proc/00_test.jpg"),
-            os.path.join(mod_path, "test/testpic_processed.jpg"),
+            os.path.join(mod_path, "img", "img_proc", "00_test.jpg"),
+            os.path.join(mod_path, "test", "testpic_processed.jpg"),
             shallow=False
         )
         self.assertTrue(success)
