+++ pfla/fcn/analyze.py
@@ -1,10 +1,10 @@
 # -*- coding: utf-8 -*-
-#------------------------------------------------------------------------------
+# -----------------------------------------------------------------------------
 #
-#  This class is responsible for calling the statistical analysis script. It 
+#  This class is responsible for calling the statistical analysis script. It
 #  will output the results and graphs to the data/ldmks directory
 #
-#------------------------------------------------------------------------------
+# -----------------------------------------------------------------------------
 import os
 import sys
 import rpy2.robjects as robjects
@@ -21,18 +21,17 @@ def main_method(mod_path):
     Parameters
     ----------
     mod_path : string
-        Path to the pfla module
+        Path to the pfla module.
 
     Returns
     -------
     None
     """
 
-    source_path = os.path.join(mod_path, "fcn/stats.R")
-    data_path = os.path.join(mod_path, "data/ldmks/")
-    source_call = "Rscript " + source_path + " " + data_path
+    source_path = os.path.join(mod_path, "fcn", "stats.R")
+    data_path = os.path.join(mod_path, "data", "ldmks")
+    source_call = "Rscript {} {}".format(source_path, data_path)
 
     rsource = robjects.r["system"]
     rpaste = robjects.r["paste"]
     rsource(rpaste(source_call))
-
