### 2018-11-29 #855 @glemaitre
I just made a PR to correct the last glitch. Nothing important regarding the paper.
