### 2018-11-29 df2a956
fixed install and usage documentation
