### 2018-11-29 #855 @arokem
@maxrousseau : would you like to include @glemaitre's additional suggestion in the version that will go along with the paper? If so, please create another archived version.
