### 2018-11-29 #855 @maxrousseau
perfect just pushed the fixed docs and will accept the pull request as soon as it passes the CI
