### 2018-11-30 7068f40
TST: adding argument when calling the tests
