### 2018-11-30 a2215f9
fixed fetcher for python 3.5 and 3.6
