+++ pfla/fcn/fetcher.py
@@ -10,4 +10,4 @@ import requests
 def data_fetch(path):
     url = 'https://github.com/maxrousseau/pfla/blob/master/pfla/data/shape_predictor_68_face_landmarks.dat?raw=true'
     r = requests.get(url, allow_redirects=True)
-    open(path, 'wb').write(r.content)
+    open(str(path), 'wb').write(r.content)
