### 2018-11-30 6146acf
version 0.2.1 and fixes for CI
