### 2018-11-30 #855 @maxrousseau
@arokem: since the PR failed the CI I would prefer to stick with the current version for publication. I will look into @glemaitre PR as per maintenance of the package independently later on.
