### 2018-11-30 f4db5fd
EHN: add cli
