### 2018-11-30 8b0b384
MAINT: move the images into data folder
