+++ pfla/fcn/fetcher.py
@@ -8,8 +8,9 @@
 import requests
 
 
-def data_fetch(path):
+def data_fetch(filename):
     url = ('https://github.com/maxrousseau/pfla/blob/master/pfla/data/'
            'shape_predictor_68_face_landmarks.dat?raw=true')
     r = requests.get(url, allow_redirects=True)
-    open(str(path), 'wb').write(r.content)
+    with open(filename, 'wb') as f:
+        f.write(r.content)
