### 2018-11-30 a924cfa
TST: test the command line and report coverage
