+++ pfla/fcn/annotate.py
@@ -6,7 +6,6 @@
 #
 # -----------------------------------------------------------------------------
 import os
-import sys
 
 import cv2
 import dlib
