+++ pfla/__init__.py
@@ -1,6 +1,9 @@
-from .fcn import img_prep
-from .fcn import face_detect
-from .fcn import annotate
-from .fcn import analyze
-from .fcn import fetcher
 from . import cli
+
+from .fcn import analyze
+from .fcn import annotate
+from .fcn import face_detect
+from .fcn import fetcher
+from .fcn import img_prep
+
+__all__ = ['analyze', 'annotate', 'cli', 'face_detect', 'fetcher', 'img_prep']
