+++ pfla/fcn/face_detect.py
@@ -7,7 +7,6 @@
 #
 # -----------------------------------------------------------------------------
 import os
-import sys
 
 import cv2
 import pandas as pd
