### 2018-11-30 5c9137e
TST: make test discoverable by pytest outside of repo
