+++ .travis.yml
@@ -4,6 +4,7 @@ group: stable
 language: python
 python:
     - 3.5
+cache: pip
 
 before_install:
     # install R
