### 2018-11-30 043e68b
MAINT: install the version from source not pypi
