### 2018-11-30 42dc42e
CI: installing R packages
