+++ .travis.yml
@@ -13,6 +13,8 @@ before_install:
     - gpg -a --export E084DAB9 | sudo apt-key add -
     - sudo apt-get update
     - sudo apt-get install r-base r-base-dev
+    - sudo R -e "install.packages(\"foreach\", repos=\"http://cran.us.r-project.org\")"
+    - sudo R -e "install.packages(\"shapes\", repos=\"http://cran.us.r-project.org\")"
     - pip3 install -r requirements.txt
 
 install: pip3 install .
