+++ .travis.yml
@@ -4,7 +4,6 @@ group: stable
 language: python
 python:
     - 3.5
-cache: pip
 
 before_install:
     # install R
