### 2018-11-30 bef9b34
CI: use conda to manage depencency
