### 2018-11-30 842aaaf
MAINT: use bash file
