+++ ci_build/install_script.sh
@@ -1,0 +1,17 @@
+if test -e $HOME/miniconda3/bin; then
+    echo "miniconda is already installed"
+else
+    wget https://repo.anaconda.com/miniconda/Miniconda3-latest-Linux-x86_64.sh -O miniconda.sh
+    chmod +x miniconda.sh
+    ./miniconda.sh -b -p $HOME/miniconda
+    export PATH=$HOME/miniconda/bin:$PATH
+    conda update --yes --quiet conda
+    conda env create -f environment.yml
+    conda activate pfla
+    mkdir shapes
+    cd shapes
+    conda skeleton cran --recursive shapes
+    conda build r-shapes
+    conda install --yes -c /home/travis/miniconda3/envs/pfla/conda-bld r-shapes
+    cd ..
+fi
