+++ ci_build/install_script.sh
@@ -5,7 +5,6 @@ if test -e $HOME/miniconda/bin; then
     if test -e 'source activate pfla'; then
         echo "The required environment exist"
         source activate pfla
-        conda install --yes -c $HOME/miniconda/envs/pfla/conda-bld r-shapes
     else
         conda env create -f environment.yml
         source activate pfla
