+++ environment.yml
@@ -9,10 +9,11 @@ dependencies:
   - progress
   - rpy2
   - opencv
-  - pytest
-  - pytest-cov
   - requests
   - r-base
   - r-foreach
+  - pytest
+  - pytest-cov
+  - flake8
   - pip:
     - imutils
