+++ .travis.yml
@@ -1,13 +1,12 @@
 os: linux
 dist: trusty
 language: python
-cache:
-  directories:
-    - $HOME/miniconda
+# cache:
+#   directories:
+#     - $HOME/miniconda
 before_install: bash ci_build/install_script.sh
 install: pip install . --force
 script:
-  - conda install --yes flake8  # temporary stuff
   - flake8 pfla
   - pytest -sv --cov=pfla pfla
   
