### 2018-11-30 #855 @glemaitre
@maxrousseau you could also make a PyPi release at the same time.
